FIX Remove an erronous "the" from a message.
The player can see invisible monsters which are bribed, charmed or hexed. When the status is lost, the monster may suddenly vanish.

This changes the message this generates from "You can no longer detect the Sigmund." to "You can no longer detect Sigmund.". The "the" is retained for non-unique monsters.